requirements.txt
awscrt==0.13.13; python_version >= "3.6"
cachetools==5.2.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7")
certifi==2022.5.18.1; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.0.12; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.5.0"
click==8.1.3; python_version >= "3.7"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
dacite==1.6.0; python_version >= "3.6"
flask==2.1.2; python_version >= "3.7"
google-api-core==2.8.2; python_version >= "3.7"
google-auth==2.9.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
google-cloud-pubsub==2.13.4; python_version >= "3.7"
googleapis-common-protos==1.56.4; python_version >= "3.7"
grpc-google-iam-v1==0.12.4; python_version >= "3.7"
grpcio-status==1.48.0; python_version >= "3.7"
grpcio==1.48.0; python_version >= "3.7"
idna==3.3; python_version >= "3.7" and python_version < "4"
itsdangerous==2.1.2; python_version >= "3.7"
jinja2==3.1.2; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
proto-plus==1.20.6; python_version >= "3.7"
protobuf==3.20.1; python_version >= "3.7"
pyasn1-modules==0.2.8; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
pyasn1==0.4.8; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") or python_full_version >= "3.6.0" and python_version >= "3.7" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7")
pycryptodome==3.14.1; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.5.0"
pysesame3==0.6.0; python_version >= "3.7" and python_version < "4.0"
pyyaml==6.0; python_version >= "3.6"
requests==2.28.0; python_version >= "3.7" and python_version < "4"
rsa==4.9; python_version >= "3.6" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7")
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_version >= "3.7" and python_full_version >= "3.6.0"
slack-sdk==3.17.2; python_full_version >= "3.6.0"
urllib3==1.26.9; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.7"
werkzeug==2.1.2; python_version >= "3.7"
